+++ code/core/src/main/java/org/betonquest/betonquest/quest/objective/damage/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Damage Objective
+ */
+package org.betonquest.betonquest.quest.objective.damage;